Moved shortcoes out into separate files
