+++ src/eleventy/shortcodes/banner.js
@@ -1,0 +1,18 @@
+module.exports = function (eleventyConfig) {
+  eleventyConfig.addPairedShortcode("banner", function (content, title) {
+    return `
+      <div class="govuk-notification-banner" role="region" aria-labelledby="govuk-notification-banner-title" data-module="govuk-notification-banner">
+        <div class="govuk-notification-banner__header">
+          <h2 class="govuk-notification-banner__title" id="govuk-notification-banner-title">
+            Important
+          </h2>
+        </div>
+        <div class="govuk-notification-banner__content">
+          <h3 class="govuk-notification-banner__heading">
+            ${title}
+          </h3>
+          ${content}</div>
+      </div>
+    `;
+  });
+}
