+++ src/eleventy/shortcodes/date-in-current-month.js
@@ -1,0 +1,6 @@
+module.exports = function (eleventyConfig) {
+  eleventyConfig.addShortcode(
+    "dateInCurrentMonth",
+    (day) => `${day}/${new Date().getMonth() + 1}/${new Date().getFullYear()}`,
+  );
+}
