+++ src/eleventy/shortcodes/version.js
@@ -1,0 +1,6 @@
+const releasePackage = require("@ministryofjustice/frontend/package.json");
+module.exports = function (eleventyConfig) {
+  eleventyConfig.addShortcode("version", function () {
+    return releasePackage.version;
+  });
+}
